enhancement: add initial routing screen for paath and kirtan with back navigation
