+++ frontend/package.json
@@ -6,8 +6,10 @@
     "@types/node": "^16.18.0",
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.0",
+    "@types/react-router-dom": "^5.3.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^7.13.0",
     "react-scripts": "5.0.1",
     "typescript": "^4.9.0",
     "web-vitals": "^2.1.0"
@@ -37,4 +39,4 @@
     ]
   },
   "proxy": "http://localhost:8000"
-} 
+}
