+++ requirements.txt
@@ -1,5 +1,5 @@
 fastapi==0.104.1
-uvicorn[standard]==0.24.0
+uvicorn==0.24.0
 # websockets==12.0  # Removed - using REST API instead
 python-multipart==0.0.6
 pydantic==2.5.0
